multi_agent_system/requirements.txt
# Multi-Agent AI System Requirements
# Core dependencies
openai>=1.3.0
anthropic>=0.7.0
chromadb>=0.4.15
sentence-transformers>=2.2.2
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.4.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.2
tenacity>=8.2.0
docker>=6.0.0

# Logging and monitoring
structlog>=23.1.0
rich>=13.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Development
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0